requirements.txt
django
django-autoslug
django-imagekit
pillow
gunicorn
faker
Unidecode==0.4.19
sympy
